Created a fgew files to store these classes,a nd changed the application controller to call the functions through the new classes
